# n8n-mcp/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: n8n-MCP API
  description: |
    The n8n-MCP API provides a comprehensive automation platform that enables users to create, manage, and execute workflows using AI-powered natural language prompts. 
    
    ## Features
    - 🤖 AI-powered workflow generation
    - 🔄 Real-time workflow execution
    - 🔐 Secure API key authentication
    - 📊 Comprehensive monitoring and analytics
    - 🧪 Testing and validation tools
    
    ## Getting Started
    1. Obtain an API key from your dashboard
    2. Review the authentication section below
    3. Try your first API call with the interactive documentation
    4. Explore our SDKs and tutorials for deeper integration
    
    ## Rate Limits
    - **Free Tier**: 1,000 requests/hour
    - **Pro Tier**: 10,000 requests/hour  
    - **Enterprise**: Custom limits available
    
    ## Support
    - 📚 [Documentation](https://docs.n8n-mcp.com)
    - 💬 [Community Discord](https://discord.gg/n8n-mcp)
    - 📧 [Support Email](mailto:support@n8n-mcp.com)
    
  version: 1.0.0
  termsOfService: https://n8n-mcp.com/terms
  contact:
    name: n8n-MCP Support
    url: https://n8n-mcp.com/support
    email: support@n8n-mcp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.n8n-mcp.com
    description: Production server
  - url: https://staging-api.n8n-mcp.com
    description: Staging server
  - url: https://sandbox-api.n8n-mcp.com
...
paths:
...
  /auth/login:
    post:
      summary: User Authentication
      description: Authenticate user and receive access token
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "secure_password"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                expires_in: 3600
                user:
                  id: "user123"
                  email: "user@example.com"
                  name: "John Doe"
                  plan: "pro"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /workflows:
    get:
      summary: List Workflows
      description: Retrieve a paginated list of workflows for the authenticated user
      operationId: listWorkflows
      tags:
        - Workflows
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Number of workflows per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: status
          in: query
          description: Filter by workflow status
          schema:
            type: string
            enum: [active, inactive, draft, error]
          example: active
        - name: search
          in: query
          description: Search workflows by name or description
          schema:
            type: string
          example: "email automation"
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
          example: "automation,email"
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
              example:
                workflows:
                  - id: "wf_123"
                    name: "Email Automation"
                    description: "Automated email campaign workflow"
                    status: "active"
                    created_at: "2024-01-01T00:00:00Z"
                    updated_at: "2024-01-07T10:30:00Z"
                    tags: ["automation", "email"]
                    nodes_count: 5
                    executions_count: 1250
                pagination:
                  page: 1
                  limit: 20
                  total: 45
                  total_pages: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create Workflow
      description: |
        Create a new workflow using either natural language description or explicit node configuration.
        
        ## Natural Language Creation
        Use the `prompt` field to describe your workflow in natural language. Our AI will generate the appropriate nodes and connections.
        
        ## Manual Creation  
        Provide explicit `nodes` and `connections` arrays for precise control over the workflow structure.
        
        ## Examples
        - "Send an email when a webhook receives data"
        - "Process CSV files and save to database"
        - "Monitor API endpoint and alert on failures"
      operationId: createWorkflow
      tags:
        - Workflows
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
            examples:
              natural_language:
                summary: Natural Language Workflow
                description: Create workflow using AI prompt
                value:
                  name: "Email Automation"
                  prompt: "Send a welcome email when someone submits the contact form on my website"
                  tags: ["automation", "email"]
                  active: true
              manual_creation:
                summary: Manual Workflow Creation
                description: Create workflow with explicit nodes
                value:
                  name: "API Data Processing"
                  description: "Fetch data from API and process it"
                  nodes:
                    - id: "webhook_1"
                      type: "webhook"
                      parameters:
                        path: "/api/data"
                        method: "POST"
                    - id: "transform_1"
                      type: "transform"
                      parameters:
                        code: "return items.map(item => ({ ...item, processed: true }));"
                  connections:
                    - source: "webhook_1"
                      target: "transform_1"
                  active: false
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
              example:
                id: "wf_456"
                name: "Email Automation"
                description: "AI-generated workflow for email automation"
                status: "active"
                nodes:
                  - id: "webhook_1"
                    type: "webhook"
                    name: "Contact Form Trigger"
                    parameters:
                      path: "/contact-form"
                      method: "POST"
                  - id: "email_1"
                    type: "email"
                    name: "Send Welcome Email"
                    parameters:
                      to: "{{ $json.email }}"
                      subject: "Welcome to our platform!"
                      template: "welcome_email"
                connections:
                  - source: "webhook_1"
                    target: "email_1"
                created_at: "2024-01-07T10:30:00Z"
                updated_at: "2024-01-07T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /workflows/{workflowId}:
    get:
      summary: Get Workflow
      description: Retrieve detailed information about a specific workflow
      operationId: getWorkflow
      tags:
        - Workflows
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          description: Unique identifier of the workflow
          schema:
            type: string
          example: "wf_123"
      responses:
        '200':
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update Workflow
      description: Update an existing workflow's configuration, nodes, or metadata
      operationId: updateWorkflow
      tags:
        - Workflows
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          description: Unique identifier of the workflow
          schema:
            type: string
          example: "wf_123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflowRequest'
            example:
              name: "Updated Email Automation"
              description: "Enhanced email automation with additional features"
              active: true
              nodes:
                - id: "webhook_1"
                  type: "webhook"
                  parameters:
                    path: "/contact-form"
                    method: "POST"
                - id: "condition_1"
                  type: "condition"
                  parameters:
                    field: "email"
                    operation: "exists"
                - id: "email_1"
                  type: "email"
                  parameters:
                    to: "{{ $json.email }}"
                    subject: "Welcome!"
              connections:
                - source: "webhook_1"
                  target: "condition_1"
                - source: "condition_1"
                  target: "email_1"
                  condition: "true"
      responses:
        '200':
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete Workflow
      description: Permanently delete a workflow and all its execution history
      operationId: deleteWorkflow
      tags:
        - Workflows
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          description: Unique identifier of the workflow
          schema:
            type: string
          example: "wf_123"
      responses:
        '204':
          description: Workflow deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workflows/{workflowId}/execute:
    post:
      summary: Execute Workflow
      description: |
        Trigger immediate execution of a workflow with optional input data.
        
        ## Execution Modes
        - **Synchronous**: Wait for completion and return results
        - **Asynchronous**: Return execution ID immediately
        
        ## Input Data
        Provide data that will be available to the workflow's first node as `$input.json`.
      operationId: executeWorkflow
      tags:
        - Workflows
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          description: Unique identifier of the workflow
          schema:
            type: string
          example: "wf_123"
        - name: mode
          in: query
          description: Execution mode
          schema:
            type: string
            enum: [sync, async]
            default: async
          example: "async"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteWorkflowRequest'
            example:
              data:
                name: "John Doe"
                email: "john@example.com"
                company: "Acme Corp"
              metadata:
                source: "api"
                campaign_id: "camp_123"
      responses:
        '200':
          description: Workflow execution completed (sync mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '202':
          description: Workflow execution started (async mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionStartedResponse'
              example:
                execution_id: "exec_789"
                status: "running"
                started_at: "2024-01-07T10:30:00Z"
                estimated_duration: 30
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workflows/{workflowId}/executions:
    get:
      summary: List Workflow Executions
      description: Retrieve execution history for a specific workflow
      operationId: listWorkflowExecutions
      tags:
        - Executions
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          description: Unique identifier of the workflow
          schema:
            type: string
          example: "wf_123"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by execution status
          schema:
            type: string
            enum: [pending, running, success, error, cancelled]
        - name: from
          in: query
          description: Start date filter (ISO 8601)
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: to
          in: query
          description: End date filter (ISO 8601)
          schema:
            type: string
            format: date-time
          example: "2024-01-07T23:59:59Z"
      responses:
        '200':
          description: List of workflow executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionListResponse'
...
  /api-keys:
    get:
      summary: List API Keys
      description: Retrieve all API keys for the authenticated user
      operationId: listApiKeys
      tags:
        - API Keys
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create API Key
      description: Generate a new API key for programmatic access
      operationId: createApiKey
      tags:
        - API Keys
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
            example:
              name: "Production Integration"
              scopes: ["read", "write"]
              expires_at: "2024-12-31T23:59:59Z"
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api-keys/{keyId}:
    delete:
      summary: Revoke API Key
      description: Permanently revoke an API key
      operationId: revokeApiKey
      tags:
        - API Keys
      security:
        - BearerAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          description: Unique identifier of the API key
          schema:
            type: string
          example: "key_123"
      responses:
        '204':
          description: API key revoked successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          description: User's password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        expires_in:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'
...
    ApiKeyListResponse:
      type: object
      properties:
        api_keys:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeySummary'
...